server/requirements.txt
pyyaml
openai
Jinja2
python-telegram-bot
pgvector
graypy
elasticsearch
databases